feat(auth-model) Breaking out auth+model to auth-model
