+++ module/auth-model/doc/config.ts
@@ -1,5 +1,5 @@
 import { InjectableFactory } from '@travetto/di';
-import { ModelAuthService, RegisteredPrincipal } from '@travetto/auth';
+import { ModelAuthService, RegisteredPrincipal } from '@travetto/auth-model';
 import { ModelCrudSupport } from '@travetto/model';
 
 import { User } from './model';
