+++ module/auth-model/src/model.ts
@@ -1,13 +1,8 @@
-// @file-if @travetto/model
 import { AppError, Util, Class } from '@travetto/base';
 import { ModelCrudSupport, ModelType, NotFoundError, OptionalId } from '@travetto/model';
 import { EnvUtil } from '@travetto/boot';
+import { AuthUtil, Principal, Authenticator, Authorizer } from '@travetto/auth';
 import { isStorageSupported } from '@travetto/model/src/internal/service/common';
-
-import { Principal } from '../types/principal';
-import { Authenticator } from '../types/authenticator';
-import { Authorizer } from '../types/authorizer';
-import { AuthUtil } from '../util';
 
 /**
  * A set of registration data
