+++ module/auth-model/test-isolated/model-file.ts
@@ -1,5 +1,3 @@
-// @with-module @travetto/model
-// @with-module @travetto/di
 import { InjectableFactory } from '@travetto/di';
 import { FileModelConfig, FileModelService } from '@travetto/model';
 import { Suite } from '@travetto/test';
