+++ module/auth-rest/src/internal/principal-target-schema.ts
@@ -1,24 +1,7 @@
-// @file-if @travetto/schema
 import { SchemaRegistry } from '@travetto/schema';
-
-import { PrincipalTarget } from '../internal/types';
-
-SchemaRegistry.register(PrincipalTarget, {
-  class: PrincipalTarget,
-});
-
-/*
-  id: string;
-  expiresAt?: Date;
-  issuedAt?: Date;
-  maxAge?: number;
-  issuer?: string;
-  details?: unknown;
-  permissions?: string[];
-*/
+import { PrincipalTarget } from '@travetto/auth/src/internal/types';
 
 const optional = (): { required: { active: false } } => ({ required: { active: false } });
-
 SchemaRegistry.registerPendingFieldConfig(PrincipalTarget, 'id', String, {});
 SchemaRegistry.registerPendingFieldConfig(PrincipalTarget, 'expiresAt', Date, optional());
 SchemaRegistry.registerPendingFieldConfig(PrincipalTarget, 'issuedAt', Date, optional());
@@ -26,3 +9,7 @@ SchemaRegistry.registerPendingFieldConfig(PrincipalTarget, 'mageAge', Number, op
 SchemaRegistry.registerPendingFieldConfig(PrincipalTarget, 'issuer', String, optional());
 SchemaRegistry.registerPendingFieldConfig(PrincipalTarget, 'permissions', [String], optional());
 SchemaRegistry.registerPendingFieldConfig(PrincipalTarget, 'details', Object, optional());
+
+SchemaRegistry.register(PrincipalTarget, {
+  class: PrincipalTarget,
+});
