+++ module/auth/index.ts
@@ -2,6 +2,3 @@ export * from './src/util';
 export * from './src/types/authenticator';
 export * from './src/types/authorizer';
 export * from './src/types/principal';
-
-// Named export needed for proxying
-export { ModelAuthService, RegisteredPrincipal } from './src/extension/model';
